Add files via upload
1) Message Reading
The first thing you have to do is to write a function that reads the contents of the files that we have.
2) Message Fusion
In this task, you have to make use of messages of two different files. In the two files, we have one number each. You have to apply a certain operation to extract our message.
3)Message Substitution
In this task, you have to substitute the message of the file for a secret message.
4)Message Comparision
In this task, you have to make use of messages from two different files. You have to compare the two messages and take only those words that appear in first message but not in second message.
5)Message Filter
In this task, you have to extract only those words from the message in the file that are of even length.
6)Message Writing
Congrats lieutenant, you have successfully deciphered all the message bits that we received. In this final task, we will combine all the message bits into a single message and write it in a file.